Here i shared a form that is you can command. These commands are Adding, Listing, Update, Delete, and searching for all datas.
First, here is an login part in my form that is configured with sql server and C#. After this form and after login you will see an main form. You can acces car datas, employee description, Mercedes branches and accessories for Mercedes. If you are logged in as a manager to my Form, you can reach to  Add, List, Update,Delete, and Search commands. In addition to these backround of my form is inspired by mercedes.